\docType{data}
\name{GolubEUPData}
\alias{GolubEUPData}
\title{A data set from Golub & Steunenberg (2007).}
\format{A data set with 3001 observations and 17 variables}
\source{
  Golub & Steunenberg (2007)
  (\url{http://hdl.handle.net/1902.1/15633})
}
\description{
  A data set from Golub & Steunenberg (2007).
}
\keyword{datasets}

